Moved manage category from Nav menu component to Manage category
