+++ DocumentApp/ClientApp/src/app/app.module.ts
@@ -23,6 +23,7 @@ import { LoadingInterceptor } from './_interceptors/loading.interceptor';
 import { NgxSpinnerModule } from 'ngx-spinner';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
+import { ManageCategoryComponent } from './manage-category/manage-category.component';
 
 @NgModule({
   declarations: [
@@ -33,7 +34,8 @@ import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
     DocEditComponent,
     NewDocComponent,
     NewCategoryComponent,
-    CategoryEditComponent
+    CategoryEditComponent,
+    ManageCategoryComponent
   ],
   imports: [
     BrowserModule.withServerTransition({ appId: 'ng-cli-universal' }),
