+++ DocumentApp/ClientApp/src/app/nav-menu/nav-menu.component.css
@@ -16,3 +16,7 @@ html {
 .box-shadow {
   box-shadow: 0 .25rem .75rem rgba(0, 0, 0, .05);
 }
+
+.nav-item{
+  display: inline-block;
+}
